feat: 분위기 축을 연속 값으로 전환
축을 유한 binary64 [-1,1]과 axisDefinitionVersion 2로 전환하고 V10으로 과거 행과 이력을 보존합니다. 원문 수치 검증, 0 정규화, 분석 영수증 v3 및 flat AI v2 계약을 적용합니다. 설정된 취향 설명 코드포인트 상한을 온보딩과 변경에 함께 적용합니다. 실제 PostgreSQL과 HTTP로 새 DB·과거 이력·경계값을 검증했으며 실제 AI 제공자 품질은 검증하지 않았습니다.

+++ src/main/java/team4/emotionmap/memory/dto/MemoryResponse.java
@@ -19,7 +19,7 @@ public record MemoryResponse(
         UUID id, UUID placeId, String content, String imageUrl,
         DistributionType distributionType, OriginKind originKind, DataOrigin dataOrigin,
         String placeLabelSnapshot, Double placeLat, Double placeLng,
-        Short crowdLevel, Short spatialFeel, Short companyFit, Short stayStyle,
+        double crowdLevel, double spatialFeel, double companyFit, double stayStyle,
         AxisSource crowdSource, AxisSource spatialSource, AxisSource companySource, AxisSource staySource,
         Short axisDefinitionVersion, AtmosphereAnalysisStatus atmosphereAnalysisStatus,
         CategoryAnalysisStatus categoryAnalysisStatus, ContentStatus contentStatus,
